Create wait list function

+++ src/app/Http/Controllers/ParticipantController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Event;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\DB;
 
 class ParticipantController extends Controller
 {
@@ -20,12 +21,16 @@ class ParticipantController extends Controller
         }
 
         $event = Event::find($eventId);
-        $event->participants()->attach(auth()->id());
-        $participants = $event->participants()->get();
+        if ($event->participants()->count() >= $event->participant_limit_number) {
+            $event->waitList()->attach(auth()->id());
+        } else {
+            $event->participants()->attach(auth()->id());
+        }
 
         return response()->json([
             'is_attended' => true,
-            'participants' => $participants,
+            'participants' => $event->participants()->get(),
+            'wait_list' => $event->waitList()->get(),
         ]);
     }
 
@@ -41,12 +46,23 @@ class ParticipantController extends Controller
         }
 
         $event = Event::find($eventId);
-        $event->participants()->detach(auth()->id());
-        $participants = $event->participants()->get();
+        if ($event->isAttended(auth()->id())) {
+            $event->participants()->detach(auth()->id());
+        }
+        // Check if there is a user on the wait_list
+        $wait_list = $event->waitList();
+        if ($wait_list->first() && !$event->isWaitListed(auth()->id())) {
+            $first_wait_list = $wait_list->first();
+            $event->participants()->attach($first_wait_list->user_id);
+            $event->waitList()->detach($first_wait_list->user_id);
+        } else {
+            $event->waitList()->detach(auth()->id());
+        }
 
         return response()->json([
             'is_attended' => false,
-            'participants' => $participants,
+            'participants' => $event->participants()->get(),
+            'wait_list' => $event->waitList()->get(),
         ]);
     }
 }
